+++ src/Tool/ActorProviderInterface.php
@@ -1,0 +1,21 @@
+<?php
+
+/*
+ * This file is part of the Doctrine Behavioral Extensions package.
+ * (c) Gediminas Morkevicius <gediminas.morkevicius@gmail.com> http://www.gediminasm.org
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Gedmo\Tool;
+
+/**
+ * Interface for a provider for an actor for extensions supporting actor/user references.
+ */
+interface ActorProviderInterface
+{
+    /**
+     * @return object|string|null
+     */
+    public function getActor();
+}